+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Security
 
 
+## [3.0.17] - 2020-04-20
+### Added
+- Added AuthUserAvailableActions::toArray() method.
+- Added helper traits with actions for using in the controller.
+
 ## [3.0.16] - 2020-04-19
 ### Added
 - Added AuthUserAvailableActions::toJson() method.
@@ -155,6 +160,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Big unification of some classes and methods.
 - Updated the Composer dependencies.
 
+[3.0.17]: https://github.com/CaliforniaMountainSnake/simple-laravel-auth-system/compare/3.0.16...3.0.17
 [3.0.16]: https://github.com/CaliforniaMountainSnake/simple-laravel-auth-system/compare/3.0.15...3.0.16
 [3.0.15]: https://github.com/CaliforniaMountainSnake/simple-laravel-auth-system/compare/3.0.14...3.0.15
 [3.0.14]: https://github.com/CaliforniaMountainSnake/simple-laravel-auth-system/compare/3.0.13...3.0.14